+++ src/home/admin.py
@@ -6,7 +6,7 @@ from adminsortable2.admin import SortableAdminMixin
 from django.contrib.admin import ModelAdmin, register
 from solo.admin import SingletonModelAdmin
 
-from home.models import Education, Experience, PersonalInfo, Project, Service, Technology
+from home.models import Education, Experience, PersonalInfo, ProcessStep, Project, Service, Technology
 
 
 @register(PersonalInfo)
@@ -25,6 +25,11 @@ class ProjectAdmin(SortableAdminMixin, ModelAdmin[Project]):  # type: ignore[mis
     filter_horizontal = ("technologies",)
     list_display = ("title", "featured", "order")
     list_editable = ("featured",)
+
+
+@register(ProcessStep)
+class ProcessStepAdmin(SortableAdminMixin, ModelAdmin[ProcessStep]):  # type: ignore[misc]
+    list_display = ("title", "order")
 
 
 @register(Service)